# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospitality Ops Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues/{venueId}/service/config:
    get:
      operationId: getServiceModeConfig
      tags: [serviceMode]
      summary: Bootstrap config for service mode — staff, inventory, equipment
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service mode bootstrap config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceBootstrapConfig"
  /venues/{venueId}/service/prep-tasks/{taskId}/complete:
    patch:
      operationId: completeServicePrepTask
      tags: [serviceMode]
      summary: Toggle a prep task complete or undo (no auth — service mode only)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completedBy:
                  type: string
                undo:
                  type: boolean
      responses:
        "200":
          description: Updated prep task
  /venues/{venueId}/service/cleaning-tasks:
    get:
      operationId: listServiceCleaningTasks
      tags: [serviceMode]
      summary: List active cleaning tasks for service mode (no auth)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of cleaning tasks
  /venues/{venueId}/service/cleaning-tasks/{taskId}/complete:
    post:
      operationId: completeServiceCleaningTask
      tags: [serviceMode]
      summary: Complete a cleaning task during service (no auth)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completedBy]
              properties:
                completedBy:
                  type: string
                notes:
                  type: string
      responses:
        "201":
          description: Cleaning log created
  /venues/{venueId}/service/temperature-log:
    post:
      operationId: createServiceTemperatureLog
      tags: [serviceMode]
      summary: Quick temperature log during service (no auth)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickTempLogInput"
      responses:
        "201":
          description: Temperature log created with pass/fail status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickTempLogResult"
  /venues/{venueId}/service/waste-log:
    post:
      operationId: createServiceWasteLog
      tags: [serviceMode]
      summary: Quick waste log during service — sets isQuick=true (no auth)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuickWasteLogInput"
      responses:
        "201":
          description: Waste log created with optional par check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuickWasteLogResult"

  # ─── SUPPLIERS ────────────────────────────────────────────────────────────
...
  /venues/{venueId}/setup-progress:
    get:
      operationId: getSetupProgress
      tags: [venues]
      summary: Get first-week setup progress and feature unlock status
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Setup progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupProgress"
components:
  schemas:
...
    SetupProgressStep:
      type: object
      required: [id, label, description, completed, action, href, unlocks]
      properties:
        id:
          type: string
        label:
          type: string
        description:
          type: string
        completed:
          type: boolean
        count:
          type: integer
        target:
          type: integer
        action:
          type: string
        href:
          type: string
        unlocks:
          type: string
    SetupProgress:
      type: object
      required: [percentComplete, stepsComplete, totalSteps, allComplete, steps]
      properties:
        percentComplete:
          type: integer
        stepsComplete:
          type: integer
        totalSteps:
          type: integer
        allComplete:
          type: boolean
        nextStep:
          $ref: "#/components/schemas/SetupProgressStep"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SetupProgressStep"

    # ─── VENUE ─────────────────────────────────────────────────────────────
...
    ServiceBootstrapConfig:
      type: object
      required: [staff, inventory, equipment]
      properties:
        staff:
          type: array
          items:
            type: object
            required: [id, name, role]
            properties:
              id:
                type: integer
              name:
                type: string
              role:
                type: string
        inventory:
          type: array
          items:
            type: object
            required: [id, name, unit]
            properties:
              id:
                type: integer
              name:
                type: string
              unit:
                type: string
              parLevel:
                type: ["string", "null"]
              currentStock:
                type: ["string", "null"]
              averageCost:
                type: ["string", "null"]
        equipment:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id:
                type: integer
              name:
                type: string
              minTemp:
                type: ["string", "null"]
              maxTemp:
                type: ["string", "null"]
        chemicals:
          type: array
          items:
            type: object
    QuickTempLogInput:
      type: object
      required: [temp]
      properties:
        temp:
          type: string
          description: Recorded temperature reading
        itemName:
          type: string
        equipmentId:
          type: integer
        checkedBy:
          type: string
    QuickTempLogResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pass, fail]
        id:
          type: integer
        venueId:
          type: integer
        equipmentId:
          type: ["integer", "null"]
        recordedTemp:
          type: string
        checkedBy:
          type: string
    QuickWasteLogInput:
      type: object
      required: [itemName, quantity, unit, reason]
      properties:
        inventoryItemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: string
        unit:
          type: string
        reason:
          type: string
        notes:
          type: string
        loggedBy:
          type: string
    QuickWasteLogResult:
      type: object
      required: [wasteLog]
      properties:
        wasteLog:
          type: object
        parCheck:
          type: object
          nullable: true
          properties:
            belowPar:
              type: boolean
            projectedStock:
              type: string
            parLevel:
              type: string
            itemName:
              type: string
...